Update settitle.txt
Added another block of settitle numbers due to the increased number of badges. Current max badge settitle is 2551, and the old version of this file stopped at 2550, so I added another block of 50 to make this file check through settitle number 2600. It should be good for another few years I think.